Linting pass on configuratt module. Rename files, get rid of * imports, get rid of circular import.
