+++ scabha/configuratt/core.py
@@ -1,125 +1,202 @@
-import os.path
 import importlib
+import os.path
 import re
-import fnmatch
+import uuid
 from collections.abc import Sequence
+from dataclasses import make_dataclass
+from typing import Any, Callable, List, Optional, Union
 
-from omegaconf.omegaconf import OmegaConf, DictConfig, ListConfig
 from omegaconf.errors import OmegaConfBaseException
-from typing import Any, List, Dict, Optional, OrderedDict, Union, Callable
+from omegaconf.omegaconf import DictConfig, ListConfig, OmegaConf
+from yaml.error import YAMLError
 
-from .common import *
+from .cache import load_cache, save_cache
+from .common import IMPLICIT_EXTENSIONS, PATH, ConfigurattError, pop_conf
 from .deps import ConfigDependencies, FailRecord
+from .helpers import _lookup_name, _scrub_subsections
 
 
-def _lookup_nameseq(name_seq: List[str], source_dict: Dict):
-    """Internal helper: looks up nested item ('a', 'b', 'c') in a nested dict
+def load(
+    path: str,
+    use_sources: Optional[List[DictConfig]] = [],
+    name: Optional[str] = None,
+    location: Optional[str] = None,
+    includes: bool = True,
+    selfrefs: bool = True,
+    include_path: str = None,
+    use_cache: bool = True,
+    no_toplevel_cache=False,
+    include_stack=[],
+    verbose: bool = False,
+):
+    """Loads config file, using a previously loaded config to resolve _use references.
 
-    Parameters
-    ----------
-    name_seq : List[str]
-        sequence of keys to look up
-    source_dict : Dict
-        nested dict
+    Args:
+        path (str): path to config file
+        use_sources (Optional[List[DictConfig]]): list of existing configs to be used to resolve "_use" references,
+                or None to disable
+        name (Optional[str]): name of this config file, used for error messages
+        location (Optional[str]): location where this config is being loaded (if not at root level)
+        includes (bool, optional): If True (default), "_include" references will be processed
+        selfrefs (bool, optional): If False, "_use" references will only be looked up in existing config.
+            If True (default), they'll also be looked up within the loaded config.
+        include_stack: list of paths which have been included. Used to catch recursive includes.
+        include_path (str, optional):
+            if set, path to each config file will be included in the section as element 'include_path'
 
-    Returns
-    -------
-    Any
-        value if found, else None
+    Returns:
+        Tuple of (conf, dependencies)
+            conf (DictConfig): config object
+            dependencies (ConfigDependencies): filenames that were _included
     """
-    source = source_dict
-    names = list(name_seq)
-    while names:
-        source = source.get(names.pop(0), None)
-        if source is None:
-            return None
-    return source
+    use_toplevel_cache = use_cache and not no_toplevel_cache
+    conf, dependencies = load_cache((path,), verbose=verbose) if use_toplevel_cache else (None, None)
+
+    if conf is None:
+        # create self:xxx resolver
+        self_namespace = dict(path=path, dirname=os.path.dirname(path), basename=os.path.basename(path))
+
+        def self_namespace_resolver(arg):
+            if arg in self_namespace:
+                return self_namespace[arg]
+            raise KeyError(f"invalid '${{self:arg}}' substitution in {path}")
+
+        OmegaConf.register_new_resolver("self", self_namespace_resolver)
+        try:
+            subconf = OmegaConf.load(path)
+            # force resolution of interpolations at this point (otherwise they happen lazily)
+            resolved = OmegaConf.to_container(subconf, resolve=True)
+            subconf = OmegaConf.create(resolved)
+        finally:
+            OmegaConf.clear_resolver("self")
+
+        name = name or os.path.basename(path)
+        dependencies = ConfigDependencies()
+        dependencies.add(path)
+        # include ourself into sources, if _use is in effect, and we've enabled selfrefs
+        if use_sources is not None and selfrefs:
+            use_sources = [subconf] + list(use_sources)
+        conf, deps = resolve_config_refs(
+            subconf,
+            pathname=path,
+            location=location,
+            name=name,
+            includes=includes,
+            use_cache=use_cache,
+            use_sources=use_sources,
+            include_path=include_path,
+            include_stack=include_stack + [path],
+        )
+        # update overall dependencies
+        dependencies.update(deps)
+
+        # # check for missing requirements
+        # dependencies.scan_requirements(conf, location, path)
+
+        if use_cache:
+            save_cache((path,), conf, dependencies, verbose=verbose)
+
+    return conf, dependencies
 
 
-def _lookup_name(name: str, *sources: List[Dict]):
-    """Internal helper: looks up a nested item ("a.b.c") in a list of dicts
+def load_nested(
+    filelist: List[str],
+    structured: Optional[DictConfig] = None,
+    typeinfo=None,
+    use_sources: Optional[List[DictConfig]] = [],
+    location: Optional[str] = None,
+    nameattr: Union[Callable, str, None] = None,
+    config_class: Optional[str] = None,
+    include_path: Optional[str] = None,
+    use_cache: bool = True,
+    verbose: bool = False,
+):
+    """Builds nested configuration from a set of YAML files corresponding to sub-sections
 
     Parameters
     ----------
-    name : str
-        section name to look up, e.g. "a.b.c"
+    conf : OmegaConf object
+        root OmegaConf object to merge content into
+    filelist : List[str]
+        list of subsection config files to load
+    schema : Optional[DictConfig]
+        schema to be applied to each file, if any
+    use_sources : Optional[List[DictConfig]]
+        list of existing configs to be used to resolve "_use" references, or None to disable
+    location : Optional[str]
+        if set, contents of files are being loaded under 'location.subsection_name'. If not set, then 'subsection_name'
+        is being loaded at root level. This is used for correctly formatting error messages and such.
+    nameattr : Union[Callable, str, None]
+        if None, subsection_name will be taken from the basename of the file. If set to a string such as 'name', will
+        set subsection_name from that field in the subsection config. If callable, will be called with the subsection
+        config object as a single argument, and must return the subsection name
+    config_class : Optional[str]
+        name of config dataclass to form (when using typeinfo), if None, then generated automatically
+    include_path : Optional[str]
+        if set, path to each config file will be included in the section as element 'include_path'
 
     Returns
     -------
-    Any
-        first matching item found
+        Tuple of (conf, dependencies)
+            conf (DictConfig): config object
+            dependencies (set): set of filenames that were _included
 
     Raises
     ------
     NameError
-        if matching item is not found
+        If subsection name is not resolved
     """
-    name_seq = name.split(".")
-    for source in sources:
-        result = _lookup_nameseq(name_seq, source)
-        if result is not None:
-            return result
-    raise ConfigurattError(f"unknown key {name}")
+    section_content, dependencies = load_cache(filelist, verbose=verbose) if use_cache else (None, None)
 
+    if section_content is None:
+        section_content = {}  # OmegaConf.create()
+        dependencies = ConfigDependencies()
 
-def _flatten_subsections(conf, depth: int = 1, sep: str = "__"):
-    """Recursively flattens subsections in a DictConfig (modifying in place)
-    A structure such as
-        a:
-            b: 1
-            c: 2
-    Becomes
-        a__b: 1
-        a__c: 2
+        for path in filelist:
+            # load file
+            subconf, deps = load(path, location=location, use_sources=use_sources, include_path=include_path)
+            dependencies.update(deps)
+            if include_path:
+                subconf[include_path] = path
 
-    Args:
-        conf (DictConfig): config to flatten
-        depth (int):       depth to which to flatten. Default is 1 level.
-        sep (str):         separator to use, default is "__"
-    """
-    subsections = [(key, value) for key, value in conf.items() if isinstance(value, DictConfig)]
-    for name, subsection in subsections:
-        pop_conf(conf, name)
-        if depth > 1:
-            _flatten_subsections(subsection, depth - 1, sep)
-        for key, value in subsection.items():
-            conf[f"{name}{sep}{key}"] = value
-
-
-def _scrub_subsections(conf: DictConfig, scrubs: Union[str, List[str]]):
-    """
-    Scrubs named subsections from a config.
-
-    Args:
-        conf (DictConfig): config to scrub
-        scrubs (Union[str, List[str]]): sections to remove (can include dots to remove nested sections)
-    """
-    if isinstance(scrubs, str):
-        scrubs = [scrubs]
-
-    for scrub in scrubs:
-        if "." in scrub:
-            name, remainder = scrub.split(".", 1)
-        else:
-            name, remainder = scrub, None
-        # apply name as pattern
-        is_pattern = "*" in name or "?" in name
-        matches = fnmatch.filter(conf.keys(), name)
-        if not matches:
-            # if no matches to pattern, it's ok, otherwise raise error
-            if is_pattern:
-                return
-            raise ConfigurattError(f"no entry matching '{name}'")
-        # recurse into or remove matching entries
-        for key in matches:
-            if remainder:
-                subconf = conf[key]
-                if type(subconf) is DictConfig:
-                    _scrub_subsections(subconf, remainder)
-                elif not is_pattern:
-                    raise ConfigurattError(f"'{name}' does not refer to a subsection")
+            # figure out section name
+            if nameattr is None:
+                name = os.path.splitext(os.path.basename(path))[0]
+            elif callable(nameattr):
+                name = nameattr(subconf)
+            elif nameattr in subconf:
+                name = subconf.get(nameattr)
             else:
-                del conf[key]
+                raise NameError(f"{path} does not contain a '{nameattr}' field")
+
+            # # resolve _use and _include statements
+            # try:
+            #     subconf = resolve_config_refs(subconf, f"{location}.{name}" if location else name, conf, subconf))
+            # except (OmegaConfBaseException, YAMLError) as exc:
+            #     raise ConfigurattError(f"config error in {path}: {exc}")
+
+            # apply schema
+            if structured is not None:
+                try:
+                    subconf = OmegaConf.merge(structured, subconf)
+                except (OmegaConfBaseException, YAMLError) as exc:
+                    raise ConfigurattError(f"schema error in {path}: {exc}")
+
+            section_content[name] = subconf
+
+        if structured is None and typeinfo is not None:
+            if config_class is None:
+                config_class = "ConfigClass_" + uuid.uuid4().hex
+            fields = [(name, typeinfo) for name in section_content.keys()]
+            datacls = make_dataclass(config_class, fields)
+            # datacls.__module__ == __name__  # for pickling
+            structured = OmegaConf.structured(datacls)
+            section_content = OmegaConf.merge(structured, section_content)
+
+        if use_cache:
+            save_cache(filelist, section_content, dependencies, verbose=verbose)
+
+    return section_content, dependencies
 
 
 def resolve_config_refs(
@@ -171,8 +248,6 @@ def resolve_config_refs(
     dependencies = ConfigDependencies()
     # self-referencing enabled if first source is ourselves
     selfrefs = use_sources and conf is use_sources[0]
-
-    from scabha.configuratt import load, PATH
 
     if isinstance(conf, DictConfig):
         ## NB: perhaps have _use and _include take effect at the point they're inserted?
@@ -235,10 +310,10 @@ def resolve_config_refs(
                             flags = {}
                             warn = optional = False
 
-                        # helper function -- finds given include file (including trying an implicit .yml or .yaml extension)
-                        # returns full name of file if found, else return None if include is optional, else
-                        # adds fail record and raises exception.
-                        # If opt=True, this is stronger than optional (no warnings raised)
+                        # helper function -- finds given include file (including trying an implicit .yml or .yaml
+                        # extension) returns full name of file if found, else return None if include is optional,
+                        # else adds fail record and raises exception. If opt=True, this is stronger than optional
+                        # (no warnings raised)
                         def find_include_file(path: str, opt: bool = False):
                             # if path already has an extension, only try the pathname itself
                             if os.path.splitext(path)[1]:
@@ -297,11 +372,13 @@ def resolve_config_refs(
                                             )
                                             if warn:
                                                 print(
-                                                    f"Warning: unable to resolve path for optional include {incl}, does {modulename} contain __init__.py?"
+                                                    f"Warning: unable to resolve path for optional include {incl}, "
+                                                    f"does {modulename} contain __init__.py?"
                                                 )
                                             continue
                                         raise ConfigurattError(
-                                            f"{errloc}: {keyword} {incl}: can't resolve path for {modulename}, does it contain __init__.py?"
+                                            f"{errloc}: {keyword} {incl}: can't resolve path for {modulename}, does "
+                                            f"it contain __init__.py?"
                                         )
                                     path = path[0]
